+++ Adj/Algebra.hs
@@ -1,1 +1,3 @@
-module Adj.Algebra where
+module Adj.Algebra (module Exports) where
+
+import Adj.Algebra.Semigroupoid as Exports