+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # 0.0.1
 * Define `Adj.Algebra` module
 * Define `Adj.Program` module
+* Define `Adj.Algebra.Semigroupoid` module